Refactor route tests to fix local test failures
