+++ backend/routes/managedenvironment/managedenviroment.go
@@ -2,7 +2,6 @@ package routes
 
 import (
 	"log"
-	"net/http"
 
 	restful "github.com/emicklei/go-restful/v3"
 )
@@ -45,32 +44,6 @@ type ManagedEnvironmentPostEntry struct {
 	URL               string `json:"url"`
 	KubeConfig        string `json:"config"`
 	KubeConfigContext string `json:"configContext"`
-}
-
-// Creating a webservice for application endpoints
-func RouteInit() {
-	ws := new(restful.WebService)
-	ws.
-		Path("/api/v1/managedenvironment").
-		Consumes(restful.MIME_JSON).
-		Produces(restful.MIME_JSON)
-
-	ws.Route(ws.GET("").To(HandleListManagedEnvironments).
-		Returns(200, "OK", "List of Managed Clusters").
-		Returns(404, "Error Occured, Not Found", nil))
-
-	ws.Route(ws.POST("").To(HandlePostManagedEnvironment).
-		Returns(200, "OK", "URL for the WebUI").
-		Returns(404, "Error Occured, Not Found", nil))
-
-	ws.Route(ws.GET("/{managedenv-id}").To(HandleGetASpecificManagementEnvironment))
-	//	ws.Route(ws.DELETE("/{managedenv-id}").To(m.deleteManagedEnviroment))
-	restful.Add(ws)
-
-	log.Print("The server is up, and listening to port 8090 on your host.")
-	server := &http.Server{Addr: ":8090"}
-	log.Fatal(server.ListenAndServe())
-
 }
 
 // Handle the GET to /api/v1/managedenvironment
