+++ backend/routes/tests/application_test.go
@@ -1,17 +1,22 @@
-package routes
+package tests
 
 import (
 	"net/http"
 	"testing"
 
-	util "github.com/redhat-appstudio/managed-gitops/backend/util"
+	"github.com/redhat-appstudio/managed-gitops/backend/routes"
+	"github.com/redhat-appstudio/managed-gitops/backend/util"
 )
 
-func TestServer(t *testing.T) {
+func TestApplication(t *testing.T) {
 	serverURL := "http://localhost:8090"
+
+	server := routes.RouteInit()
 	go func() {
-		RunRestfulCurlyRouterServer()
+		server.ListenAndServe()
 	}()
+	defer server.Close()
+
 	if err := util.WaitForServerUp(serverURL); err != nil {
 		t.Errorf("%v", err)
 	}
